qa: restructure redis related tests to prevent side-effects
